refactor injecting functions into structure
