+++ crates/sdk/src/signer/mod.rs
@@ -1,6 +1,6 @@
 pub mod core;
 pub mod error;
-mod wtns_parser;
+mod wtns_injector;
 
 pub use core::{Signer, SignerTrait};
 pub use error::SignerError;
